Fix an issue where image bounds weren't being precisely respected and block ContextMenu on image when in draw mode

+++ TileShop.UI/Features/Graphics/GraphicsEditorViewModel.cs
@@ -391,7 +391,7 @@ public sealed partial class GraphicsEditorViewModel : ResourceEditorBaseViewMode
 
     private void ReloadImage() => _imageAdapter.Render();
 
-    public bool ContainsPoint(int x, int y)
+    public bool ContainsPoint(double x, double y)
     {
         return x >= 0 && y >= 0 && x < WorkingArranger.ArrangerPixelSize.Width && y < WorkingArranger.ArrangerPixelSize.Height;
     }
